refactoring and color button css setting
